REBASEME: rm req/res type from constructor, use valid typehints
